requirements.txt
asyncio==3.4.3
certifi==2019.3.9
chardet==3.0.4
idna==2.8
praw==6.1.1
prawcore==1.0.1
requests==2.21.0
six==1.15.0
update-checker==0.16
urllib3==1.24.1
websocket-client==0.56.0
selenium==3.141.0
discord.py==1.3.4
tornado==5.1.1
pymongo==3.9.0
postgres==3.0.0
psycopg2-binary==2.8.6
psycopg2-pool==1.1
google==3.0.0
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
pytz==2020.5